module authorization done

+++ controllers/course-Management/module.js
@@ -2,10 +2,12 @@ import { findCourseById, updateCourseById } from "../../repositories/course.js";
 import {
   addModule,
   deleteModuleById,
+  findAllModules,
   findModuleById,
   findModulesByCourseId,
   updateModuleById,
 } from "../../repositories/module.js";
+import { findUserById } from "../../repositories/user.js";
 
 export const createModule = async (req, res) => {
   try {
@@ -62,13 +64,16 @@ export const getSingleModule = async (req, res) => {
 export const getModuleByCourse = async (req, res) => {
   try {
     const { id } = req.params;
+
     const findCourse = await findCourseById(id);
     if (!findCourse)
       return res.status(404).json({ message: "course not found" });
+
     const findModule = await findModulesByCourseId(findCourse._id);
     if (findModule.length < 1) {
       return res.status(404).json({ message: "no modules found" });
     }
+
     res.status(200).json(findModule);
   } catch (err) {
     res
@@ -80,11 +85,34 @@ export const getModuleByCourse = async (req, res) => {
 export const updateModule = async (req, res) => {
   try {
     const { id } = req.params;
+    const userId = req.user.id;
     const body = req.body;
-    const updatedModule = await updateModuleById(id, body);
-    if (!updatedModule) {
-      return res.status(404).json({ message: "no module found to update" });
+
+    const findUser = await findUserById(userId);
+    if (!findUser) {
+      return res.status(404).json({ message: "user not found, Invalid Token" });
     }
+
+    const findModule = await findModuleById(id);
+    if (!findModule) {
+      return res
+        .status(404)
+        .json({ message: "no module found with the provided ID to update" });
+    }
+
+    if (
+      !findUser?.courses?.some((courseId) =>
+        courseId.equals(findModule.courseId),
+      ) &&
+      findUser.role !== "admin"
+    ) {
+      return res
+        .status(403)
+        .json({ message: "Forbidden. You are not the owner of this course." });
+    }
+
+    const updatedModule = await updateModuleById(id, body);
+
     res.status(200).json(updatedModule);
   } catch (err) {
     res
@@ -96,12 +124,32 @@ export const updateModule = async (req, res) => {
 export const deleteModule = async (req, res) => {
   try {
     const { id } = req.params;
+    const userId = req.user.id;
+
+    const findUser = await findUserById(userId);
+    if (!findUser) {
+      return res.status(404).json({ message: "user not found, Invalid Token" });
+    }
+
     const findModule = await findModuleById(id);
-    await updateCourseById(findModule.courseId, { $pull: { module: id } });
-    const deleteModule = await deleteModuleById(id);
-    if (!deleteModule) {
+    if (!findModule) {
       return res.status(404).json({ message: "no module found to delete" });
     }
+
+    if (
+      !findUser?.courses?.some((courseId) =>
+        courseId.equals(findModule.courseId),
+      ) &&
+      findUser.role !== "admin"
+    ) {
+      return res
+        .status(403)
+        .json({ message: "Forbidden. You are not the owner of this course." });
+    }
+
+    await updateCourseById(findModule.courseId, { $pull: { module: id } });
+    await deleteModuleById(id);
+
     res.status(200).json({ message: "deleted successfully" });
   } catch (err) {
     res
